pyproject.toml
[project]
name = "graphrag-agent"
version = "0.1.0"
description = "Multi-tenant GraphRAG agent with LangChain & LangGraph"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "GraphRAG Team"}
]

dependencies = [
    # Core
    "langchain>=0.3.0",
    "langgraph>=0.2.0",
    "langchain-groq>=0.2.0",
    "langchain-community>=0.3.0",
    "langchain-huggingface>=0.1.0",
    "langchain-text-splitters>=0.3.0",
    
    # Graph Database
    "neo4j>=5.0.0",
    
    # Vector Store (PostgreSQL + pgvector)
    "asyncpg>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "pgvector>=0.3.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.13.0",
    
    # Object Storage
    "minio>=7.2.0",
    
    # Embeddings
    "sentence-transformers>=2.2.0",
    
    # API
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.9",
    
    # HTTP Client
    "httpx>=0.27.0",
    
    # Utilities
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    
    # Community Detection
    "networkx>=3.0",
    "python-louvain>=0.16",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "black>=24.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
